move `pkg/sexp` to `pkg/util/sexp`
This just moves the `sexp` package to be within the `util` package,
given that this most accurately reflects its usage.
