+++ pkg/air/lisp.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 
 	sc "github.com/consensys/go-corset/pkg/schema"
-	"github.com/consensys/go-corset/pkg/sexp"
+	"github.com/consensys/go-corset/pkg/util/sexp"
 )
 
 // Lisp converts this schema element into a simple S-Expression, for example
