+++ pkg/corset/expression.go
@@ -5,9 +5,9 @@ import (
 	"math/big"
 	"reflect"
 
-	"github.com/consensys/go-corset/pkg/sexp"
 	tr "github.com/consensys/go-corset/pkg/trace"
 	"github.com/consensys/go-corset/pkg/util"
+	"github.com/consensys/go-corset/pkg/util/sexp"
 )
 
 // Expr represents an arbitrary expression over the columns of a given context
