+++ pkg/schema/assignment/computed_column.go
@@ -2,9 +2,9 @@ package assignment
 
 import (
 	sc "github.com/consensys/go-corset/pkg/schema"
-	"github.com/consensys/go-corset/pkg/sexp"
 	"github.com/consensys/go-corset/pkg/trace"
 	"github.com/consensys/go-corset/pkg/util"
+	"github.com/consensys/go-corset/pkg/util/sexp"
 )
 
 // ComputedColumn describes a column whose values are computed on-demand, rather
